feat : 차트 앱 추가 후 수정 완료
